requirements.txt
flask
flask-sqlalchemy
google-generativeai
playwright
beautifulsoup4
python-dotenv
requests
flask-cors
